libstore: extract S3 multipart upload into S3CompatBinaryCacheStore
The GCS XML API implements the S3 multipart-upload protocol verbatim
so the upcoming GCS store would otherwise duplicate lines of S3BinaryCacheStore.

This commit pulls the multipart lifecycle, Content-MD5 hashing, compression dispatch and
the multipart-* settings into an S3CompatBinaryCacheStore[Config] base.

Backends supply only prepareRequest() (setupForS3/setupForGCS), backendName() for diagnostics.
