+++ src/libstore/include/nix/store/meson.build
@@ -86,6 +86,7 @@ headers = [ config_pub_h ] + files(
   'remote-store.hh',
   'restricted-store.hh',
   's3-binary-cache-store.hh',
+  's3-compat-binary-cache-store.hh',
   's3-url.hh',
   'serve-protocol-connection.hh',
   'serve-protocol-impl.hh',
